Add files via upload
Diabetic retinopathy Detection Using Deep Learning algorithms:
	This project is a categorical classification of different stages of Diabetic Retinopathy Stages, where using different deep learning techniques like CNN (Convolutional Neural Network), Densenet, VGG16, VGG19(Visual Geometry Group), ResNetInceptionV2, Depth separable Model.
This project is a comparative study of above different Deep Learning algorithms and how they classified the stages and getting chance of DR and explored how they trained the models over a dataset of APTOS 2019 from Kaggle. DR is trained for real-time image quality assessment, lesion detection and grading using 466,247 fundus images from 121,342 patients with diabetes. Evaluation is performed on a local dataset with 200,136 fundus images from 52,004 patients and three external datasets with a total of 209,322 images.
Our model scored 74% ,98%, 72%, 84%, 95%, 60%, 52% in accuracy while predicting labels for the validation dataset, which is a great performance, and it was competent to identify relevant patterns across all the classes in the dataset. Among them Densenet and Oversampling of VGG19 Got the Highest accuracy to identify and classify the Diabetic retinopathy stages